FIX/ENH: resolve circular imports, make widget discovery even more automatic
