+++ pcdswidgets/builder/ui/positioner_row_tc_interlock.ui
@@ -66,24 +66,12 @@
        <number>5</number>
       </property>
       <item>
-       <widget class="PyDMEmbeddedDisplay" name="PyDMEmbeddedDisplay">
+       <widget class="PositionerRowWidget" name="PositionerRowWidget">
         <property name="toolTip">
          <string/>
         </property>
-        <property name="macros" stdset="0">
-         <string>{}</string>
-        </property>
-        <property name="filename" stdset="0">
-         <string>./positioner_row.ui</string>
-        </property>
-        <property name="loadWhenShown" stdset="0">
-         <bool>true</bool>
-        </property>
-        <property name="disconnectWhenHidden" stdset="0">
-         <bool>true</bool>
-        </property>
-        <property name="followSymlinks" stdset="0">
-         <bool>false</bool>
+        <property name="motor" stdset="0">
+         <string>${MOTOR}</string>
         </property>
        </widget>
       </item>
@@ -120,28 +108,28 @@
           <property name="toolTip">
            <string/>
           </property>
-          <property name="precision" stdset="0">
+          <property name="precision">
            <number>1</number>
           </property>
-          <property name="showUnits" stdset="0">
+          <property name="showUnits">
            <bool>false</bool>
           </property>
-          <property name="precisionFromPV" stdset="0">
+          <property name="precisionFromPV">
            <bool>false</bool>
           </property>
-          <property name="alarmSensitiveContent" stdset="0">
+          <property name="alarmSensitiveContent">
            <bool>false</bool>
           </property>
-          <property name="alarmSensitiveBorder" stdset="0">
+          <property name="alarmSensitiveBorder">
            <bool>true</bool>
           </property>
-          <property name="PyDMToolTip" stdset="0">
+          <property name="PyDMToolTip">
            <string/>
           </property>
-          <property name="channel" stdset="0">
+          <property name="channel">
            <string>ca://${MOTOR}:ILOCK:TC_TEMP_RBV</string>
           </property>
-          <property name="enableRichText" stdset="0">
+          <property name="enableRichText">
            <bool>false</bool>
           </property>
          </widget>
@@ -180,51 +168,51 @@
           <property name="toolTip">
            <string/>
           </property>
-          <property name="alarmSensitiveContent" stdset="0">
+          <property name="alarmSensitiveContent">
            <bool>false</bool>
           </property>
-          <property name="alarmSensitiveBorder" stdset="0">
+          <property name="alarmSensitiveBorder">
            <bool>true</bool>
           </property>
-          <property name="PyDMToolTip" stdset="0">
+          <property name="PyDMToolTip">
            <string/>
           </property>
-          <property name="channel" stdset="0">
+          <property name="channel">
            <string>ca://${MOTOR}:ILOCK:ACTIVE_RBV</string>
           </property>
-          <property name="onColor" stdset="0">
+          <property name="onColor">
            <color>
             <red>255</red>
             <green>0</green>
             <blue>0</blue>
            </color>
           </property>
-          <property name="offColor" stdset="0">
+          <property name="offColor">
            <color>
             <red>0</red>
             <green>255</green>
             <blue>0</blue>
            </color>
           </property>
-          <property name="showLabels" stdset="0">
+          <property name="showLabels">
            <bool>false</bool>
           </property>
-          <property name="bigEndian" stdset="0">
+          <property name="bigEndian">
            <bool>false</bool>
           </property>
-          <property name="circles" stdset="0">
+          <property name="circles">
            <bool>false</bool>
           </property>
-          <property name="labelPosition" stdset="0">
+          <property name="labelPosition">
            <enum>QTabWidget::West</enum>
           </property>
-          <property name="numBits" stdset="0">
+          <property name="numBits">
            <number>1</number>
           </property>
-          <property name="shift" stdset="0">
+          <property name="shift">
            <number>0</number>
           </property>
-          <property name="labels" stdset="0">
+          <property name="labels">
            <stringlist>
             <string>Bit 0</string>
            </stringlist>
@@ -263,9 +251,9 @@
    <header>pydm.widgets.byte</header>
   </customwidget>
   <customwidget>
-   <class>PyDMEmbeddedDisplay</class>
-   <extends>QFrame</extends>
-   <header>pydm.widgets.embedded_display</header>
+   <class>PositionerRowWidget</class>
+   <extends>QWidget</extends>
+   <header>pcdswidgets.motion.positioner_row_widget</header>
   </customwidget>
  </customwidgets>
  <resources/>
