+++ pcdswidgets/entrypoint_widgets.py
@@ -5,30 +5,32 @@ section in pyproject.toml
 python -m pcdswidgets.entrypoint_widgets
 """
 
+import importlib
 import inspect
+import pkgutil
 from pathlib import Path
-from typing import cast
+from types import ModuleType
+from typing import Iterator, cast
 
 import tomlkit as tk
 import tomlkit.items as tki
 
-import pcdswidgets.eps_byteindicator
-import pcdswidgets.motion
-import pcdswidgets.table
-import pcdswidgets.vacuum
+SKIP_WIDGETS = [
+    "PCDSSymbolBase",
+]
 
-INCLUDE_MODULES = [
-    pcdswidgets.eps_byteindicator,
-    pcdswidgets.motion,
-    pcdswidgets.table,
-    pcdswidgets.vacuum,
+SKIP_MODULES = [
+    ".tests",
+    ".demo",
 ]
 
 
 def main():
     key_val: list[tuple[str, str]] = []
-    for module in INCLUDE_MODULES:
+    for module in iter_submodules():
         for name, obj in inspect.getmembers(module, inspect.isclass):
+            if name in SKIP_WIDGETS:
+                continue
             if hasattr(obj, "_qt_designer_"):
                 key_val.append((name, f"{obj.__module__}:{name}"))
     key_val.sort()
@@ -48,6 +50,20 @@ def main():
 
     with open(pyproj, "w") as fd:
         tk.dump(toml_doc, fd)
+
+
+def iter_submodules(package: str = "pcdswidgets") -> Iterator[ModuleType]:
+    """Recursively yield all submodules of a package."""
+    if any(mod in package for mod in SKIP_MODULES):
+        return
+    module = importlib.import_module(package)
+    yield module
+    try:
+        for _, modname, _ in pkgutil.walk_packages(module.__path__, module.__name__ + "."):
+            if "__main__" not in modname:
+                yield from iter_submodules(modname)
+    except AttributeError:
+        ...
 
 
 if __name__ == "__main__":
