--- pcdswidgets/motion/__init__.py
@@ -1,1 +1,0 @@
-from .motors import *  # noqa: F403
