+++ pcdswidgets/motion/positioner_row_tc_interlock_widget.py
@@ -1,0 +1,8 @@
+from pcdswidgets.builder.ui.positioner_row_tc_interlock_base import Positioner_Row_Tc_InterlockWidgetBase
+
+
+class PositionerRowTcInterlockWidget(Positioner_Row_Tc_InterlockWidgetBase):
+    _qt_designer_ = {
+        "group": "PCDS Motion",
+        "is_container": False,
+    }
