+++ pcdswidgets/motion/positioner_row_widget.py
@@ -1,0 +1,8 @@
+from pcdswidgets.builder.ui.positioner_row_base import Positioner_RowWidgetBase
+
+
+class PositionerRowWidget(Positioner_RowWidgetBase):
+    _qt_designer_ = {
+        "group": "PCDS Motion",
+        "is_container": False,
+    }
