+++ pcdswidgets/motion/positioner_widget.py
@@ -1,0 +1,8 @@
+from pcdswidgets.builder.ui.positioner_base import PositionerWidgetBase
+
+
+class PositionerWidget(PositionerWidgetBase):
+    _qt_designer_ = {
+        "group": "PCDS Motion",
+        "is_container": False,
+    }
