+++ pcdswidgets/motion/smaract_open_loop_widget.py
@@ -1,0 +1,8 @@
+from pcdswidgets.builder.ui.smaract_open_loop_base import Smaract_Open_LoopWidgetBase
+
+
+class SmaractOpenLoopWidget(Smaract_Open_LoopWidgetBase):
+    _qt_designer_ = {
+        "group": "PCDS Motion",
+        "is_container": False,
+    }
