+++ pyproject.toml
@@ -44,15 +44,15 @@ NeedleValve = "pcdswidgets.vacuum.valves:NeedleValve"
 PneumaticValve = "pcdswidgets.vacuum.valves:PneumaticValve"
 PneumaticValveDA = "pcdswidgets.vacuum.valves:PneumaticValveDA"
 PneumaticValveNO = "pcdswidgets.vacuum.valves:PneumaticValveNO"
-PositionerRowTcInterlockWidget = "pcdswidgets.motion.motors:PositionerRowTcInterlockWidget"
-PositionerRowWidget = "pcdswidgets.motion.motors:PositionerRowWidget"
-PositionerWidget = "pcdswidgets.motion.motors:PositionerWidget"
+PositionerRowTcInterlockWidget = "pcdswidgets.motion.positioner_row_tc_interlock_widget:PositionerRowTcInterlockWidget"
+PositionerRowWidget = "pcdswidgets.motion.positioner_row_widget:PositionerRowWidget"
+PositionerWidget = "pcdswidgets.motion.positioner_widget:PositionerWidget"
 ProportionalValve = "pcdswidgets.vacuum.valves:ProportionalValve"
 RGA = "pcdswidgets.vacuum.others:RGA"
 RightAngleManualValve = "pcdswidgets.vacuum.valves:RightAngleManualValve"
 RoughGauge = "pcdswidgets.vacuum.gauges:RoughGauge"
 ScrollPump = "pcdswidgets.vacuum.pumps:ScrollPump"
-SmaractOpenLoopWidget = "pcdswidgets.motion.motors:SmaractOpenLoopWidget"
+SmaractOpenLoopWidget = "pcdswidgets.motion.smaract_open_loop_widget:SmaractOpenLoopWidget"
 TurboPump = "pcdswidgets.vacuum.pumps:TurboPump"
 
 [project.optional-dependencies]
